app/agents/meta_agents/strategic_meta_agent.py
from typing import Any, Dict, List
from app.models.agent.goal import Goal
from app.agents.meta_agents.meta_agents import MetaAgent
from app.tools.reasoning_engine import ReasoningEngine
from app.tools.ontology_reasoner import OntologyReasoner
from app.models.knowledge.knowledge_base import KnowledgeBase
from app.tools.llm_manager import LLMManager
from app.models.knowledge.ontology.ontology import Ontology
from app.services.agent_service import AgentService
from app.services.goal_service import GoalService
from app.services.plan_service import PlanService
import asyncio

class StrategicMetaAgent(MetaAgent):
    def __init__(self, agent_id: str, name: str, api_key: str, llm_service: Any, agent_communication_service: Any):
        super().__init__(
            agent_id=agent_id,
            name=name,
            api_key=api_key,
            llm_service=llm_service,
            agent_communication_service=agent_communication_service
        )
        self.agent_type = "strategic_meta"
        self.knowledge_base = KnowledgeBase()
        self.reasoning_engine = ReasoningEngine(self.knowledge_base, api_key)
        self.llm_manager = LLMManager()
        self.ontology = Ontology()  # Initialize with your strategic ontology
        self.ontology_reasoner = OntologyReasoner(self.llm_manager, self.ontology)
        self.agent_service = AgentService()
        self.goal_service = GoalService()
        self.plan_service = PlanService()

        self._init_beliefs()
        self._init_desires()
        self._init_goals()
        self._init_plans()

    def _init_beliefs(self):
        self.add_belief("Strategic planning is crucial for long-term MAS success")
        self.add_belief("Resource allocation significantly impacts system performance")

    def _init_desires(self):
        self.add_desire("Develop and maintain effective long-term strategies", priority=10)
        self.add_desire("Optimize resource allocation across the MAS", priority=9)

    def _init_goals(self):
        self.add_goal("Analyze system state and environment", priority=9)
        self.add_goal("Generate strategic goals", priority=8)
        self.add_goal("Allocate resources efficiently", priority=8)
        self.add_goal("Monitor and adjust strategies", priority=7)

    def _init_plans(self):
        self.create_plan(
            self.goals[0].id,
            [
                "Gather system performance data",
                "Analyze external factors",
                "Identify strategic opportunities",
                "Formulate long-term objectives"
            ]
        )

    async def reason(self):
        print("Starting reasoning process for strategic planning")
        
        current_beliefs = [belief.to_dict() for belief in self.beliefs]
        updated_beliefs = await self.reasoning_engine.reason({"beliefs": current_beliefs})
        
        for belief in updated_beliefs.get("beliefs", []):
            self.add_belief(belief["content"])

        new_knowledge = await self.ontology_reasoner.infer_new_knowledge()
        
        for concept in new_knowledge.get("new_concepts", []):
            self.add_belief(f"New strategic concept: {concept['name']} - {concept['description']}")
        
        for relationship in new_knowledge.get("new_relationships", []):
            self.add_belief(f"New strategic relationship: {relationship['name']} between {relationship['domain']} and {relationship['range']}")

        strategy_query = "What are the key strategic considerations for this MAS?"
        strategic_considerations = await self.ontology_reasoner.answer_query(strategy_query)
        self.add_belief(f"Key strategic considerations: {strategic_considerations}")

        current_beliefs = [belief.to_dict() for belief in self.beliefs]
        new_desires = await self.reasoning_engine.generate_desires(current_beliefs)
        
        for desire in new_desires:
            self.add_desire(desire["description"], desire["priority"])

        print("Reasoning process for strategic planning completed")

    async def plan(self):
        print("Starting planning process for strategic management")
        for goal in self.goals:
            if goal.description == "Generate strategic goals":
                self.create_plan(
                    goal.id,
                    [
                        "Analyze current system state",
                        "Identify long-term objectives",
                        "Formulate strategic goals",
                        "Prioritize goals",
                        "Develop action plans"
                    ]
                )
        
        current_state = self.get_current_state()
        optimized_plan = await self.reasoning_engine.reason_and_plan(self.goals[0].description, current_state)
        
        if optimized_plan.get("plan"):
            self.update_plan(self.goals[0].id, optimized_plan["plan"].steps)
        
        print("Planning process for strategic management completed")

    async def execute(self):
        print("Starting execution process for strategic management")
        for plan in self.plans:
            for task in plan.steps:
                if task.status == "pending":
                    print(f"Executing task: {task.description}")
                    try:
                        execution_result = await self.reasoning_engine.simulate_action(task.description, self.get_current_state())
                        self.update_task_status(task.id, "completed")
                        
                        for key, value in execution_result.items():
                            self.add_belief(f"Strategic task result - {key}: {value}")
                    except Exception as e:
                        print(f"Error executing strategic task {task.description}: {str(e)}")
                        self.update_task_status(task.id, "failed")
        print("Execution process for strategic management completed")

    def get_current_state(self) -> Dict[str, Any]:
        return {
            "beliefs": [belief.to_dict() for belief in self.beliefs],
            "desires": [desire.to_dict() for desire in self.desires],
            "goals": [goal.to_dict() for goal in self.goals],
            "plans": [plan.to_dict() for plan in self.plans],
        }

    async def run(self):
        while True:
            await self.reason()
            await self.plan()
            await self.execute()
            await asyncio.sleep(3600)  # Sleep for an hour, adjust as needed for strategic planning frequency

    async def analyze_system_state(self) -> Dict[str, Any]:
        return {
            "resources": await self.agent_service.get_available_resources(),
            "current_goals": await self.goal_service.get_active_goals(),
            "ongoing_plans": await self.plan_service.get_active_plans(),
            "performance_metrics": await self.agent_service.get_system_performance()
        }

    async def generate_strategic_goals(self) -> List[Goal]:
        system_state = await self.analyze_system_state()
        strategic_goals = await self.goal_service.generate_strategic_goals(system_state)
        
        # Use the reasoning engine to refine and prioritize goals
        refined_goals = await self.reasoning_engine.reason_about_goals(strategic_goals)
        
        return refined_goals

    async def allocate_resources(self, goals: List[Goal]) -> Dict[str, Any]:
        resource_allocation = {}
        available_resources = await self.agent_service.get_available_resources()
        
        for goal in goals:
            required_resources = await self.goal_service.estimate_required_resources(goal)
            allocated = await self.agent_service.allocate_resources(goal, required_resources, available_resources)
            resource_allocation[goal.id] = allocated
            available_resources = await self.agent_service.update_available_resources(available_resources, allocated)
        
        # Use the reasoning engine to optimize resource allocation
        optimized_allocation = await self.reasoning_engine.optimize_resource_allocation(resource_allocation, available_resources)
        
        return optimized_allocation

    async def monitor_progress(self, goals: List[Goal]) -> Dict[str, Any]:
        progress_report = {}
        for goal in goals:
            goal_progress = await self.goal_service.get_goal_progress(goal.id)
            associated_plans = await self.plan_service.get_plans_for_goal(goal.id)
            plan_progress = [await self.plan_service.get_plan_progress(plan.id) for plan in associated_plans]
            progress_report[goal.id] = {
                "goal_progress": goal_progress,
                "plan_progress": plan_progress
            }
        
        # Use the reasoning engine to analyze progress and suggest adjustments
        analysis_and_suggestions = await self.reasoning_engine.analyze_progress(progress_report)
        
        return analysis_and_suggestions

    async def adjust_strategy(self, progress: Dict[str, Any]) -> None:
        for goal_id, goal_progress in progress.items():
            if goal_progress["goal_progress"] < 0.5:  # If progress is less than 50%
                await self._reallocate_resources(goal_id)
                await self._revise_goal(goal_id)

    async def _reallocate_resources(self, goal_id: str) -> None:
        goal = await self.goal_service.get_goal(goal_id)
        current_allocation = await self.agent_service.get_resource_allocation(goal_id)
        new_allocation = await self.reasoning_engine.suggest_resource_reallocation(goal, current_allocation)
        await self.agent_service.update_resource_allocation(goal_id, new_allocation)

    async def _revise_goal(self, goal_id: str) -> None:
        goal = await self.goal_service.get_goal(goal_id)
        revised_goal = await self.reasoning_engine.revise_goal(goal)
        await self.goal_service.update_goal(revised_goal)
